Move tests into bootserver tests
